softpwm module is tested and complete

+++ firmware/src/SoftPWM.h
@@ -9,14 +9,17 @@
  * If the developer decides to change the PBCLK3 frequency or timer4's
  * prescaler, these macros must be updated to the appropriate values.
  * 
- * Due to hardware constraints, maximum PWM frequency is limited to 10kHz.
- * Default frequency is 5kHz. Maximum number of pins is 16 but this can be
+ * Due to hardware constraints, maximum PWM frequency is limited to 5kHz.
+ * Default frequency is 1kHz. Maximum number of pins is 16 but this can be
  * changed if desired. MICRO_NUM_PINS should be set with the number of
  * available pins on the chosen microcontroller.
  * 
  * When enabling a pin for the first time, its duty cycle should be set first:
  *      SoftPWM_PinSetDuty(<pin>, <duty>);
  *      SoftPWM_PinEnable(<pin>);
+ * 
+ * When a pin's duty cycle is changed, the wave will complete its current cycle
+ * before the changes take affect.
  *
  * Created on August 6, 2021, 3:08 PM
  */
@@ -60,7 +63,7 @@ extern "C" {
     void SoftPWM_PinDisableAll(void);
     void SoftPWM_PinSetDuty(GPIO_PIN pin, uint8_t duty);
     bool SoftPWM_SetFrequency(uint16_t freqHz);
-    uint32_t SoftPWM_GetFrequency(void);
+    uint16_t SoftPWM_GetFrequency(void);
     // **** END GLOBAL FUNCTION PROTOTYPES ****
     // -------------------------------------------------------------------------
 #ifdef	__cplusplus
